pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rinsed-snowflake-client"
version = "0.1.0"
description = "Python client for WashU Carwash Rinsed CRM data in Snowflake"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "snowflake-connector-python>=3.0,<4",
    "pydantic>=2.10,<3",
    "pandas>=2.0,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "python-dotenv>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/rinsed_snowflake_client"]
